+++ rcjaRegistration/publicapi/views.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework.decorators import action
 from django.db.models.deletion import ProtectedError
 
-from common.apiPermissions import ReadOnly, CmsSecretPermission
+from common.apiPermissions import ReadOnly
 
 from .serializers import StateSerializer, EventSerializer, SummaryEventSerializer
 
@@ -47,7 +47,7 @@ class StateViewSet(viewsets.ReadOnlyModelViewSet, NestedSerializerActionMinxin):
 
     def upcomingEventsQueryset(self, abbreviation, includeGlobal):
         return self.eventsBaseQueryset(abbreviation, includeGlobal).filter(startDate__gte=datetime.datetime.today()).order_by('startDate')
-
+    
     def pastEventsQueryset(self, abbreviation, includeGlobal):
         return self.eventsBaseQueryset(abbreviation, includeGlobal).filter(startDate__lt=datetime.datetime.today()).order_by('-startDate')
 
@@ -113,41 +113,3 @@ class StateViewSet(viewsets.ReadOnlyModelViewSet, NestedSerializerActionMinxin):
     #     state = get_object_or_404(State, pk=pk, typeWebsite=True)
     #     queryset = CommitteeMember.objects.filter(state=state).order_by('id')
     #     return self.nestedSerializer(queryset, CommitteeMemberSerializer)
-
-
-class CMSIntegrationViewSet(viewsets.ViewSet):
-    permission_classes = [CmsSecretPermission]
-
-    @action(detail=False, methods=["POST"])
-    def linkEvent(self, request):
-        """
-        Expects:
-          - eventId (the ID of the event in Rego)
-          - cmsEventId (the ID of the event in the CMS)
-
-        Updates the Event's cmsEventId field (only for competitions)
-        """
-        event_id = request.data.get("eventId")
-        cms_event_id = request.data.get("cmsEventId")
-
-        if not (event_id and cms_event_id):
-            return Response(
-                {"detail": "Both 'eventId' and 'cmsEventId' are required."},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        event = get_object_or_404(Event, pk=event_id)
-
-        if event.eventType != "competition":
-            return Response(
-                {"detail": "cmsEventId can only be set for competitions."},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        event.cmsEventId = cms_event_id
-        event.save()
-
-        return Response(
-            {"detail": f"Successfully linked '{cms_event_id}' to '{event}'."},
-            status=status.HTTP_200_OK
-        )
